Update .gitignore
Now api_info.json is ignored, and api_info_real.json is not. Now real data can be stored in api_info.json, which will not be stored in the online repository, and will not be overwritten every time we pull the repo to the production server. If someone wants to try the application for testing purposes, can redirect the code to the fake file or rename the file.