Fix for error when there are more than 100 active students
Due to api paging, only the first 100 hits of every quesry was handled. This was not a problem so far, but now there are 105 students and there were reports that some of them don't get the scheduled emails. This change fixes the problem, it queries second page as well if there are more than 100 students with a status.
Note: this solution should be generalized.