Error fix in handlewaitinglist
Earlier if the student applied to a course, and the currently advertised courses changed, it could cause a problem that the system thought the student applied to one of the new courses. If there was a new place on the new course (and there are free spots on new courses) the student got an email that they got enrolled. Now this problem is fixed and always the original course is considered.